+++ src/lib.rs
@@ -4,9 +4,9 @@ mod plugin;
 mod sync_wrapped;
 mod utils;
 
-pub use crate::audio_player::AudioListener;
-pub use crate::audio_player::AudioSource;
-pub use crate::audio_player::AudioSourcePlayer;
+pub use crate::audio_player::audio_listener::AudioListener;
+pub use crate::audio_player::audio_source::AudioSource;
+pub use crate::audio_player::AudioPlayer;
 pub use crate::one_shot_player::OneShotPlayer;
 pub use crate::plugin::FmodPlugin;
 pub use crate::sync_wrapped::event_instance::EventInstance;
